Introduce Text implementations #1461
