+++ src/test/java/org/cactoos/text/TextOfTest.java
@@ -28,28 +28,15 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
-import java.text.MessageFormat;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetDateTime;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.util.Calendar;
-import java.util.Locale;
-import java.util.TimeZone;
 import org.cactoos.bytes.BytesOf;
 import org.cactoos.io.InputOf;
 import org.cactoos.iterator.IteratorOfChars;
 import org.hamcrest.Matchers;
 import org.hamcrest.core.IsEqual;
 import org.hamcrest.core.IsNot;
-import org.hamcrest.core.IsNull;
 import org.junit.jupiter.api.Test;
 import org.llorllale.cactoos.matchers.Assertion;
 import org.llorllale.cactoos.matchers.HasString;
-import org.llorllale.cactoos.matchers.IsBlank;
 import org.llorllale.cactoos.matchers.IsText;
 
 /**
@@ -328,242 +315,6 @@ final class TextOfTest {
                 new IOException("").getStackTrace()
             ),
             new HasString("org.cactoos.text.TextOfTest")
-        ).affirm();
-    }
-
-    @Test
-    void readsLocalDateFormattedWithFormatString() {
-        final LocalDate date = LocalDate.of(2017, 12, 13);
-        new Assertion<>(
-            "Can't format a LocalDate with format.",
-            new TextOf(date, "yyyy-MM-dd HH:mm:ss"),
-            new IsText("2017-12-13 00:00:00")
-        ).affirm();
-    }
-
-    @Test
-    void readsLocalDateFormattedWithFormatStringWithLocale() {
-        final LocalDate date = LocalDate.of(2017, 12, 13);
-        new Assertion<>(
-            "Can't format a LocalDate with format using locale.",
-            new TextOf(
-                date, "yyyy MMM dd. HH.mm.ss", Locale.FRENCH
-            ),
-            new IsText("2017 déc. 13. 00.00.00")
-        ).affirm();
-    }
-
-    @Test
-    void readsLocalDateFormattedAsIsoDateTime() throws IOException {
-        final LocalDate date = LocalDate.of(2017, 12, 13);
-        new Assertion<>(
-            "Can't format a LocalDate with default/ISO format.",
-            new TextOf(date),
-            new IsText(
-                MessageFormat.format(
-                "2017-12-13T00:00:00{0}",
-                date.atTime(LocalTime.MIN).atZone(ZoneId.systemDefault())
-                    .getOffset().toString()
-                )
-            )
-        ).affirm();
-    }
-
-    @Test
-    void readsCurrentLocalDateAsText() throws Exception {
-        new Assertion<>(
-            "Can't format a LocalDate with ISO format.",
-            new TextOf(LocalDate.now()).asString(),
-            new IsNot<>(new IsBlank())
-        ).affirm();
-    }
-
-    @Test
-    void localDateTimeFormattedAsIsoDateTime() {
-        final LocalDateTime date = LocalDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17
-        );
-        new Assertion<>(
-            "Can't format a LocalDateTime with default/ISO format.",
-            new TextOf(date),
-            new IsText(
-                MessageFormat.format(
-                    "2017-12-13T14:15:16.000000017{0}",
-                    date.atZone(ZoneId.systemDefault()).getOffset().toString()
-                )
-            )
-        ).affirm();
-    }
-
-    @Test
-    void localDateTimeFormattedWithFormatString() {
-        final LocalDateTime date = LocalDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17
-        );
-        new Assertion<>(
-            "Can't format a LocalDateTime with format.",
-            new TextOf(date, "yyyy-MM-dd HH:mm:ss"),
-            new IsText("2017-12-13 14:15:16")
-        ).affirm();
-    }
-
-    @Test
-    void localDateTimeFormattedWithFormatStringWithLocale() {
-        final LocalDateTime date = LocalDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17
-        );
-        new Assertion<>(
-            "Can't format a LocalDateTime with format using locale.",
-            new TextOf(
-                date, "yyyy MMM dd. HH.mm.ss", Locale.FRENCH
-            ),
-            new IsText("2017 déc. 13. 14.15.16")
-        ).affirm();
-    }
-
-    @Test
-    void currentLocalDateTimeAsText() throws Exception {
-        new Assertion<>(
-            "Can't format a LocalDateTime with ISO format.",
-            new TextOf(LocalDateTime.now()).asString(),
-            new IsNot<>(new IsNull<>())
-        ).affirm();
-    }
-
-    @Test
-    void dateFormattedUsingIsoFormatter() {
-        final Calendar calendar =
-            Calendar.getInstance(TimeZone.getDefault());
-        calendar.set(2017, Calendar.DECEMBER, 13, 14, 15, 16);
-        calendar.set(Calendar.MILLISECOND, 17);
-        final ZoneOffset offset = calendar.getTimeZone().toZoneId()
-            .getRules().getOffset(calendar.toInstant());
-        new Assertion<>(
-            "Can't format a java.util.Date with ISO format.",
-            new TextOf(calendar.getTime()),
-            new IsText("2017-12-13T14:15:16.017" + offset)
-        ).affirm();
-    }
-
-    @Test
-    void dateFormattedUsingCustomFormat()  {
-        final Calendar calendar =
-            Calendar.getInstance(TimeZone.getDefault());
-        calendar.set(2017, Calendar.DECEMBER, 13, 14, 15, 16);
-        new Assertion<>(
-            "Can't format a java.util.Date with custom format.",
-            new TextOf(
-                calendar.getTime(), "yyyy MM dd hh:mm:ss"
-            ),
-            new IsText("2017 12 13 02:15:16")
-        ).affirm();
-    }
-
-    @Test
-    void dateFormattedUsingCustomFormatDifferentLocale() {
-        final Calendar calendar =
-            Calendar.getInstance(TimeZone.getDefault());
-        calendar.set(2017, Calendar.DECEMBER, 13, 14, 15, 16);
-        new Assertion<>(
-            "Can't format a java.util.Date with custom format.",
-            new TextOf(
-                calendar.getTime(), "yyyy MMM dd hh:mm:ss", Locale.ITALIAN
-            ),
-            new IsText("2017 dic 13 02:15:16")
-        ).affirm();
-    }
-
-    @Test
-    void offsetDateTimeFormattedAsIsoDateTime() {
-        final OffsetDateTime date = OffsetDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneOffset.ofHours(1)
-        );
-        new Assertion<>(
-            "Can't format a OffsetDateTime with default/ISO format.",
-            new TextOf(date),
-            new IsText("2017-12-13T14:15:16.000000017+01:00")
-        ).affirm();
-    }
-
-    @Test
-    void offsetDateTimeFormattedWithFormatString() {
-        final OffsetDateTime date = OffsetDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneOffset.ofHours(1)
-        );
-        new Assertion<>(
-            "Can't format a OffsetDateTime with format.",
-            new TextOf(date, "yyyy-MM-dd HH:mm:ss"),
-            new IsText("2017-12-13 14:15:16")
-        ).affirm();
-    }
-
-    @Test
-    void offsetDateTimeFormattedWithFormatStringWithLocale() {
-        final OffsetDateTime date = OffsetDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneOffset.ofHours(1)
-        );
-        new Assertion<>(
-            "Can't format a OffsetDateTime with format using locale.",
-            new TextOf(
-                date, "yyyy MMM dd. HH.mm.ss", Locale.FRENCH
-            ),
-            new IsText("2017 déc. 13. 14.15.16")
-        ).affirm();
-    }
-
-    @Test
-    void currentOffsetDateTimeAsText() throws Exception {
-        new Assertion<>(
-            "Can't format a OffsetDateTime with ISO format.",
-            new TextOf(OffsetDateTime.now()).asString(),
-            new IsNot<>(new IsNull<>())
-        ).affirm();
-    }
-
-    @Test
-    void zonedDateTimeFormattedAsIsoDateTime() {
-        final ZonedDateTime date = ZonedDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneId.of("Europe/Berlin")
-        );
-        new Assertion<>(
-            "Can't format a ZonedDateTime with default/ISO format.",
-            new TextOf(date),
-            new IsText("2017-12-13T14:15:16.000000017+01:00")
-        ).affirm();
-    }
-
-    @Test
-    void zonedDateTimeFormattedWithFormatString() {
-        final ZonedDateTime date = ZonedDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneId.of("Europe/Berlin")
-        );
-        new Assertion<>(
-            "Can't format a ZonedDateTime with format.",
-            new TextOf(date, "yyyy-MM-dd HH:mm:ss"),
-            new IsText("2017-12-13 14:15:16")
-        ).affirm();
-    }
-
-    @Test
-    void zonedDateTimeFormattedWithFormatStringWithLocale() {
-        final ZonedDateTime date = ZonedDateTime.of(
-            2017, 12, 13, 14, 15, 16, 17, ZoneId.of("Europe/Berlin")
-        );
-        new Assertion<>(
-            "Can't format a ZonedDateTime with format using locale.",
-            new TextOf(
-                date, "yyyy MMM dd. HH.mm.ss", Locale.FRENCH
-            ),
-            new IsText("2017 déc. 13. 14.15.16")
-        ).affirm();
-    }
-
-    @Test
-    void currentZonedDateTimeAsText() throws Exception {
-        new Assertion<>(
-            "Can't format a ZonedDateTime with ISO format.",
-            new TextOf(ZonedDateTime.now()).asString(),
-            new IsNot<>(new IsNull<>())
         ).affirm();
     }
